requirements.txt
asgiref==3.8.1
cffi==1.17.1
colorama==0.4.6
coverage==7.7.0
cryptography==44.0.2
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2025.1
